% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Calculate bootstrap confidence intervals for XGBP}
\usage{
bootstrap(xgbp_out, boot_iter = 100, ci_level = 0.95, verbose = TRUE)
}
\arguments{
\item{xgbp_out}{A \code{xgbp} object returned by the \code{\link{xgbp}} function}

\item{boot_iter}{Number of bootstrap iterations used to estimate non-parametric
confidence intervals. Defaults to \code{100}}

\item{ci_level}{Level of confidence intervals. Defaults to \code{0.95} (95\% CI)}

\item{verbose}{Should the function report messages along the estimation? Defaults to \code{TRUE}}
}
\description{
\code{bootstrap} uses sample with replacement from provided survey data to calculate
non-parametric confidence intervals for the \code{xgbp} function
}
\section{Parallelization}{
\code{get_estimates} uses \code{furrr}'s \code{\link[furrr:future_map]{furrr::future_map()}} to perform bootstrap iterations.
By default, computation is done sequentially, but users can take advantage of parallelism
by declaring a plan using \code{\link[future:plan]{future::plan()}}. See below for an example of setting \code{multisession}
parallelism.
}

