% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{census06}
\alias{census06}
\title{US Census 06 data}
\format{
An object of class \code{data.frame} with 2940 rows and 9 columns.
}
\usage{
census06
}
\description{
US Census 06 data
}
\keyword{datasets}
