% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poststratification.R
\name{get_estimates}
\alias{get_estimates}
\title{Get estimates for XGBP}
\usage{
get_estimates(xgbp_out, ...)
}
\arguments{
\item{xgbp_out}{A \code{xgbp} object returned by the \code{\link{xgbp}} function}

\item{...}{Group-level covars to aggregate results (leave blank for
sample-level estimates)}
}
\description{
\code{get_estimates} aggregates and returns estimates from a XGBP object.
}
\examples{
\dontrun{
get_estimates(xgbp_out)
}

}
