% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning.R
\name{tune_xgbp}
\alias{tune_xgbp}
\title{Tune XGBoost model}
\usage{
tune_xgbp(
  survey,
  ...,
  dep_var = NULL,
  nrounds = 100,
  nthread = 1,
  n_iter = 25,
  seed = NULL
)
}
\arguments{
\item{survey}{A \code{tibble} created by the \code{\link{xgbp}} function (passed internally).}

\item{...}{Individual and group level covariates used in the poststratification. All
variables must be included in the \code{survey} and in the \code{census} and passed unquoted to
the function call}

\item{dep_var}{Dependent variable. Must be \code{character} or \code{factor}}

\item{nrounds}{Number of trees (rounds) used in to train the model. Defaults to \code{100}}

\item{nthread}{Number of htreads used in the computation. Defaults to \code{1}, but users are
encourage to increase this number to speed up computations (the limit is the actual number
of threads available at your computer)}

\item{n_iter}{When \code{tune = TRUE}, this indicates how many samples to draw
during gridsearch to use. Defaults to \code{30}.}

\item{seed}{A seed for replication. Defaults to \code{NULL}}
}
\value{
A list with two elements:
\itemize{
\item A list with the best parameters selected during gridsearch
\item A vector with the optimal number of trees selected during gridsearch
}
}
\description{
\code{tune_xgbp} tunes the parameters of XGBoost to improve first stage predicitons
and avid overfitting in the XGBP.
}
