% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgbp.R
\name{xgbp}
\alias{xgbp}
\title{Extreme Gradient Boosting with Poststratification}
\usage{
xgbp(
  survey,
  census,
  census_count,
  ...,
  dep_var = NULL,
  seed = NULL,
  tune = TRUE,
  params = NULL,
  nrounds = 100,
  n_iter = 25,
  nthread = 1,
  verbose = TRUE
)
}
\arguments{
\item{survey}{A survey sample containing the variables to use in the poststratification.
Must be a \code{data.frame} or a \code{tibble}}

\item{census}{Census data to use in the poststratification. Must be a \code{data.frame} or
\code{tibble} containing the same variables, with the same categories, as the \code{survey} object}

\item{census_count}{\code{numeric} variable in the \code{census} object indicanting the raw number
or proportion of individuals in a given stratum}

\item{...}{Individual and group level covariates used in the poststratification. All
variables must be included in the \code{survey} and in the \code{census} and passed unquoted to
the function call}

\item{dep_var}{Dependent variable. Must be \code{character} or \code{factor}}

\item{seed}{A seed for replication. Defaults to \code{NULL}}

\item{tune}{Should the XGBP tune the parameters with randomized grid search? Defaults to \code{TRUE}, in which
case \code{params} argument is ignored}

\item{params}{A list of parameters to be passed to xgboost function}

\item{nrounds}{Number of trees (rounds) used in to train the model. Defaults to \code{100}}

\item{n_iter}{When \code{tune = TRUE}, this indicates how many samples to draw
during gridsearch to use. Defaults to \code{30}.}

\item{nthread}{Number of htreads used in the computation. Defaults to \code{1}, but users are
encourage to increase this number to speed up computations (the limit is the actual number
of threads available at your computer)}

\item{verbose}{Should the function report messages along the estimation? Defaults to \code{TRUE}}
}
\value{
A list of class \code{xgbp} with the following items
\itemize{
\item \code{estimates} -- A \code{tibble} containing raw estimates by strata
\item \code{model} -- The trained \code{xgboost} model
\item \code{data} -- GBP datamatrix used to train the model
\item \code{nrounds} -- Number of rounds used to train the model
\item \code{census} -- Census data used to poststratify results
\item \code{census_count} -- Variable in the \code{census} object indicanting the raw number
or proportion of individuals in a given stratum
\item \code{covars_matrix} -- GBP matrix with covars used to train the model
\item \code{dep_var} -- Dependent variable (target)
\item \code{seed} -- Seed used to reproduce results
}
}
\description{
The package's main function, \code{xgbp} poststratify a survey response from a sample
using Extreme Gradient Boosting (XGB). Dependent variables can be both binomial
or multinomial and resulting estimates can be aggregated by the full sample or
by any group used in the estimation.
}
\examples{
\dontrun{
# General use case
ps <- xgbp(survey, census, var1, var2, dep_var = Y)
}

}
