% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgbp.R
\name{xgbp}
\alias{xgbp}
\title{Extreme Gradient Boosting with Poststratification}
\usage{
xgbp(
  survey,
  census,
  census_count,
  ...,
  dep_var = NULL,
  seed = 44,
  max.depth = 6,
  eta = 0.3,
  nrounds = 100,
  subsample = 0.95,
  min_child_weight = 0.9,
  early_stopping_rounds = 20,
  nthread = 1
)
}
\arguments{
\item{survey}{A survey sample containing the variables to use in the poststratification.
Must be a \code{data.frame} or a \code{tibble}}

\item{census}{Census data to use in the poststratification. Must be a \code{data.frame} or
\code{tibble} containing the same variables, with the same categories, as the \code{survey} object}

\item{census_count}{\code{numeric} variable in the \code{census} object indicanting the raw number
or proportion of individuals in a given stratum}

\item{...}{Individual and group level covariates used in the poststratification. All
variables must be included in the \code{survey} and in the \code{census} and passed unquoted to
the function call}

\item{dep_var}{Dependent variable. Must be \code{character} or \code{factor}}

\item{seed}{A seed for replication. Defaults to \code{44}}

\item{max.depth}{XGB's max depth. Indicates the maximum depth of a tree. Defaults to \code{6}.}

\item{eta}{XGB's step size shrinkage used in update to prevents overfitting. Defaults to \code{0.3}}

\item{nrounds}{XGB's number of rounds for boosting. Defaults to \code{100}.}

\item{subsample}{XGB's subsample ratio of the training instances. Defaults to \code{0.95}}

\item{min_child_weight}{XGB's minimum sum of instance weight (hessian) needed in a child.
Defaults to \code{0.9}}

\item{early_stopping_rounds}{XGB's early stopping. Validation metric needs to improve at
least once in every early_stopping_rounds round(s) to continue training. Defaults to \code{20}}

\item{nthread}{Number of htreads used in the computation. Defaults to \code{1}, but users are
encourage to increase this number to speed up computations (the limit is the actual number
of threads available at your computer)}
}
\description{
The package's main function, \code{xgbp} poststratify a survey response from a sample
using Extreme Gradient Boosting (XGB). Dependent variables can be both binomial
or multinomial and resulting estimates can be aggregate by the whole sample or
by any group used in the estimation.
}
\examples{
\dontrun{
x <- xgbp(survey, census, var1, var2, dep_var = Y)
}

}
